Add recipients

+++ src/main/java/com/poneres/portal/signatures/SignwellSignatureService.java
@@ -1,6 +1,8 @@
 package com.poneres.portal.signatures;
 
+import com.poneres.portal.agreements.SignatureRecipient;
 import com.poneres.portal.helpers.RestApiService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
@@ -8,7 +10,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.*;
+import java.util.stream.IntStream;
 
+@Slf4j
 @Service
 public class SignwellSignatureService implements SignatureService {
 
@@ -22,21 +26,18 @@ public class SignwellSignatureService implements SignatureService {
     private String apiUrl;
 
     @Override
-    public String create(String name, Boolean draft, Boolean withSignaturePage, String base64File) {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("X-Api-Key", apiKey);
-
+    public String create(String name, Boolean draft, Boolean withSignaturePage, String base64File, List<SignatureRecipient> recipients) {
         Map<String, Object> body = new HashMap<>();
         body.put("test_mode", true);
         body.put("draft", draft);
         body.put("with_signature_page", withSignaturePage);
 
-        List<Object> recipients = new ArrayList<>();
-        recipients.add(new HashMap<>() {{
-            put("id", 1);
-            put("email", "evangelos.poneres@gmail.com");
-        }});
-        body.put("recipients", recipients);
+        body.put("recipients", IntStream.range(0, recipients.size())
+                .mapToObj(i -> new HashMap<>() {{
+                    put("id", i);
+                    put("email", recipients.get(i).getEmail());
+                    put("name", recipients.get(i).getName());
+                }}).toList());
 
         List<Object> files = new ArrayList<>();
         files.add(new HashMap<>() {{
@@ -45,15 +46,18 @@ public class SignwellSignatureService implements SignatureService {
         }});
         body.put("files", files);
 
-        Map<String, Object> response = restApiService.post(apiUrl + "documents", headers, body, HashMap.class);
+        Map<String, Object> response = restApiService.post(apiUrl + "documents", headers(), body, HashMap.class);
         return (String) response.get("id");
     }
 
     @Override
     public void delete(String ssdId) {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("X-Api-Key", apiKey);
+        if (ssdId == null) {
+            log.info("NULL ssdId was supplied");
+            return;
+        }
 
+        Map<String, String> headers = headers();
         Map<String, Object> document = restApiService.get(apiUrl + "documents/" + ssdId, headers, HashMap.class);
         if ((Boolean) document.get("test_mode")) {
             restApiService.delete(apiUrl + "documents/" + ssdId, headers);
@@ -70,17 +74,15 @@ public class SignwellSignatureService implements SignatureService {
     @Override
     public String status(String ssdId) {
         if (ssdId == null) {
-            return null;
+            log.info("NULL ssdId was supplied");
+            return "None";
         }
 
-        Map<String, String> headers = new HashMap<>();
-        headers.put("X-Api-Key", apiKey);
-
         try {
-            Map<String, Object> response = restApiService.get(apiUrl + "documents/" + ssdId, headers, HashMap.class);
+            Map<String, Object> response = restApiService.get(apiUrl + "documents/" + ssdId, headers(), HashMap.class);
             return (String) response.get("status");
         } catch (Exception ex) {
-            // Do nothing;
+            log.info("Error retrieving document status", ex);
         }
 
         return "None";
@@ -89,27 +91,27 @@ public class SignwellSignatureService implements SignatureService {
     @Override
     public String fileUrl(String ssdId) {
         if (ssdId == null) {
-            return null;
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "NULL ssdId was supplied");
         }
 
-        Map<String, String> headers = new HashMap<>();
-        headers.put("X-Api-Key", apiKey);
-
         String uri = apiUrl + "documents/" + ssdId + "/completed_pdf?url_only=true";
-        Map<String, Object> response = restApiService.get(uri, headers, HashMap.class);
+        Map<String, Object> response = restApiService.get(uri, headers(), HashMap.class);
         return (String) response.get("file_url");
     }
 
     @Override
     public void sendReminder(String ssdId) {
         if (ssdId == null) {
-            return;
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "NULL ssdId was supplied");
         }
 
+        restApiService.post(apiUrl + "documents/" + ssdId + "/remind", headers(), Collections.emptyMap());
+    }
+
+    private Map<String, String> headers() {
         Map<String, String> headers = new HashMap<>();
         headers.put("X-Api-Key", apiKey);
-
-        restApiService.post(apiUrl + "documents/" + ssdId + "/remind", headers, Collections.emptyMap());
+        return headers;
     }
 
 }
